requirements.txt
robotframework>=7.0
robotframework-seleniumlibrary>=6.3.0
selenium>=4.18.1
webdriver-manager>=4.0.1
robotframework-datadriver>=1.11.2
robotframework-requests>=0.9.7
